+++ src/c++/lib/alignment/GlobalAligner.hh
@@ -42,12 +42,6 @@ struct GlobalAligner : public SingleRefAlignerBase<ScoreType>
         SingleRefAlignerBase<ScoreType>(scores)
     {}
 
-    GlobalAligner(
-            const AlignmentScores<ScoreType>& scores,
-            const AlignmentScores<ScoreType>& contigFilterScores) :
-            SingleRefAlignerBase<ScoreType>(scores,contigFilterScores)
-    {}
-
     /// returns alignment path of query to reference
     template <typename SymIter>
     void